💄 Use responsive field rows on the security page
Collapse the two password cards into a single "Authentication" section
containing one responsive field row, matching the pattern used on the
general and profile settings pages.

Settings rows show state and expose an explicit affordance to change it,
rather than presenting the input inline. The password forms move into
dialogs triggered from the row, following DeleteAccountDialog.
