+++ apps/web/public/locales/en/app.json
@@ -123,6 +123,7 @@
   "changeNameInfo": "This will not affect any votes you have already made.",
   "changePassword": "Change password",
   "changePasswordDescription": "Update your password to keep your account secure",
+  "changePasswordSettingHint": "Update your password to keep your account secure.",
   "changeRole": "Change role",
   "charactersRemaining": "{count, plural, one {# character remaining} other {# characters remaining}}",
   "checkForConflicts": "Toggle which calendars to check for conflicts",
@@ -675,6 +676,8 @@
   "seatsUpdatedAlertTitle": "Seats updated",
   "seatUsage": "{usedSeats} of {totalSeats} seats used",
   "security": "Security",
+  "securityAuthentication": "Authentication",
+  "securityAuthenticationDescription": "Manage how you sign in to your account",
   "securityDescription": "Manage your account security and password settings",
   "selectPlan": "Select plan:",
   "selectPotentialDates": "Select potential dates or times for your event",
@@ -686,6 +689,7 @@
   "setMaxAttendees": "Set max attendees",
   "setPassword": "Set password",
   "setPasswordSectionDescription": "Your account doesn't have a password yet. Set one to sign in with your email address.",
+  "setPasswordSettingHint": "Your account doesn't have a password yet. Set one to sign in with your email address.",
   "setPasswordTitle": "Set password",
   "settingCalendarSelection": "Setting calendar selection...",
   "settingDefaultCalendar": "Setting default calendar...",
